And actually use the new worker
