+++ bag_verifier/src/test/scala/uk/ac/wellcome/platform/archive/bagverifier/fixtures/WorkerServiceFixture.scala
@@ -5,7 +5,7 @@ import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.messaging.fixtures.SNS.Topic
 import uk.ac.wellcome.messaging.fixtures.SQS.Queue
 import uk.ac.wellcome.messaging.fixtures.worker.AlpakkaSQSWorkerFixtures
-import uk.ac.wellcome.platform.archive.bagverifier.services.{NewBagVerifierWorker, Verifier}
+import uk.ac.wellcome.platform.archive.bagverifier.services.{BagVerifierWorker, Verifier}
 import uk.ac.wellcome.platform.archive.common.fixtures.{MonitoringClientFixture, OperationFixtures}
 import uk.ac.wellcome.platform.archive.common.storage.services.StorageManifestService
 
@@ -19,7 +19,7 @@ trait WorkerServiceFixture
                                outgoingTopic: Topic,
                                queue: Queue =
                                  Queue("fixture", arn = "arn::fixture"))(
-    testWith: TestWith[NewBagVerifierWorker, R]): R =
+    testWith: TestWith[BagVerifierWorker, R]): R =
     withMonitoringClient { implicit monitoringClient =>
       withActorSystem { implicit actorSystem =>
         withMaterializer(actorSystem) { implicit materializer =>
@@ -31,7 +31,7 @@ trait WorkerServiceFixture
           withIngestUpdater("verification", ingestTopic) { ingestUpdater =>
             withOutgoingPublisher("verification", outgoingTopic) {
               outgoingPublisher =>
-                val service = new NewBagVerifierWorker(
+                val service = new BagVerifierWorker(
                   alpakkaSQSWorkerConfig = createAlpakkaSQSWorkerConfig(queue),
                   ingestUpdater = ingestUpdater,
                   outgoingPublisher = outgoingPublisher,
